Refactored ProjectProfilePropertiesList component
