+++ lib/databases/board.ts
@@ -1,6 +1,6 @@
 import { v4 } from 'uuid';
 
-import type { SelectedVariables } from 'components/common/DatabaseEditor/components/viewSidebar/viewSourceOptions/components/ProposalSourcePropertiesSelectModal';
+import type { SelectedProperties } from 'components/common/DatabaseEditor/components/viewSidebar/viewSourceOptions/components/ProposalSourceProperties/ProposalSourcePropertiesDialog';
 import type { UIBlockWithDetails } from 'lib/databases/block';
 import { createBlock } from 'lib/databases/block';
 import type { PageContent } from 'lib/prosemirror/interfaces';
@@ -90,7 +90,7 @@ export type GoogleFormSourceData = {
 
 export type BoardFields = {
   icon: string;
-  selectedVariables?: SelectedVariables;
+  selectedProperties?: SelectedProperties;
   description: PageContent;
   showDescription?: boolean;
   isTemplate?: boolean;
